.github/PULL_REQUEST_TEMPLATE: remove +1 footer
The footer has the following problems:
- It's not consistently added in all PRs, for example backports or those
where the template is not used. Inconsistency in a voting skews the
results.
- It's arguably not effective, because people use reactions naturally to
interact with PRs anyway. As long as there are multiple different
reaction types, what does it mean that one PR has 20 :+1:, but another
has 22 :tada:? The latter won't show up in the ranking for :+1:, so at
the minimum confusing. Same for :heart:.
- It's not useful to determine priority as in "higher number of votes
should be merged". A PR's quality should be determined independently of
votes - when it's good, it will be merged. The currently most upvoted PR
has serious issues, so this invites drive-by questions like "Why isn't
this merged, given the number of upvotes?".

(cherry picked from commit c89c90dfa137316f3cf662783b4795d8b9897237)