python3Packages.pytestCheckHook: reduce parentheses for -m and -k options
The `-k` and `-m` options take expressions which can be constructed with
`and`, `or`, `not` and - parentheses. However, an expression like `not
(mark)` seems to be interpreted as the parentheses belonging to the mark
and not to the expression.

Thus, this commit reduces the parentheses used in expression building.
The only case left is when both enabled and disabled items are joined.
Example:

```
(a or b) and not c and not d
```