gams: move assert behind meta.license check
This now errors with the "allow unfree" error first - and only when
that's passed it asserts the specific `licenseFile` option.

This ensures that CI will not produce an eval warning for the assert
once we turn that on.

(cherry picked from commit 136dcc5f93936a795854bd9ddcbdb5e55f3708e6)