budgie-desktop-with-plugins: Populate GI_TYPELIB_PATH from plugins
Things happen to work because budgie-desktop depends on libpeas,
which propagates gobject-introspection, and so far we only package
one Python budgie plugin (budgie-media-player-applet) and that
plugin does not require deps that are not available in budgie-desktop.

Let's actually handle GI_TYPELIB_PATH in -with-plugins wrapper so
the paths actually come from plugins's buildInputs.