gccWithoutTargetLibc: move assert to meta.badPlatforms
An assert can't be caught by CI for generation of attrpaths during Eval.
An error reported via meta.badPlatforms can be.

The alternative of returning null in the CI case, similar to the commits
before, is not helpful, because it only causes failed builds or eval
downstream - when this dependency is passed on and used as `null`.