hyprland: fix eval failure on darwin
Darwin is not a supported platform, but the `meta.platform` check never
kicks in. That's because there is a `throw` in `useMoldLinker`
triggering for darwin. This, among many others, causes CI to require
hacky workarounds for evaluation.

By not adding this for darwin, eval up to `meta.platforms` will succeed
and the package then be filtered out correctly.

(cherry picked from commit 321430385fb368d3f521cda167ee539e239b701d)