hare: fix eval without enableCrossCompilation
In a `a -> b` construct, when the first condition `a` doesn't apply, the
overall result will be `true` - but this immediately fired the `throw`
in these cases. Thus `enableCrossCompilation = false` resulted in
immediate eval errors.

This also fixes eval for CI on darwin, where the assert should not
trigger, but instead let `meta.badPlatforms` be evaluated. This then
throws an "unsupported system" error, which can be caught properly by
CI.