lash: delay assert behind eval of `meta`
This allows CI to evaluate `meta.platforms` to see that this package is
not supported on darwin. Working around the assert before that requires
much more hackery.

I'm not 100% sure whether the assert is needed, but it doesn't hurt to
keep it in this place.

(cherry picked from commit 07c441b8fee23fde28347554c71df4cb6c3e7342)