tracy-wayland: drop darwin as supported platform
Wayland support is only possible on Linux. Instead of the assert, which
can't be caught by CI properly, do this via the list of supported
platforms in meta.platforms. This works much better for CI.

I did not touch the `withGtkFileSelector` assert, because it doesn't
trigger in CI - it only does when overriding this argument manually.