buildRustCrate: add a way to compile integration tests using CARGO_BIN_EXE_...
when *compiling* integration tests, cargo exports the following
variable:
CARGO_BIN_EXE_<name> — The absolute path to a binary target’s executable. This is only set when building an integration test or benchmark. This may be used with the env macro to find the executable to run for testing purposes. The <name> is the name of the binary target, exactly as-is. For example, CARGO_BIN_EXE_my-program for a binary named my-program. Binaries are automatically built when the test is built, unless the binary has required features that are not enabled.

(excerpt of the documentation)

so buildRustCrate in buildTests mode must take as argument the crate
built in normal mode, so that it may point to it.