top-level/aliases: abort evaluation on check error
The `aliases.nix` file is full of "expected errors", which the CI check
for "does nixpkgs with aliases eval?" *must* ignore. However, this also
hides an unexpected error, the check which prevents shadowing of
existing attributes in all-packages.nix.

By elevating this check to use `abort`, this will break the CI check for
sure.