python3Packages.imageio: fix build, enable most tests
With 624897829bf52e44f4d1277200357e2ae7244a55, the new parenthesization logic exposes a coding defect in `imageio` where they do a string test vs. `"not needs_internet"` instead of something parsed.

Instead of fixing that, we can actually enable almost all of the tests in this module with a command-line flag which points to a FOD of https://github.com/imageio/test_images repository.

There are a few more tests that need to be disabled as well; I opened a PR against imageio to fix these.

I also removed the `libgl-path.patch` in favor of a simple `--replace-fail`.