workflows/build: avoid downloading from cachix
This avoids downloading results from cachix, when they don't need to be
rebuild, which just wastes time and resources.

To speed up the installation of nix-build-cached, we checkout the pinned
nixpkgs instead of downloading it as a tarball. This is a bit faster and
can potentially save from worktree checkouts in the future.