multipass: 1.15.1 -> 1.16.0
This commit bumps Canonical Multipass to the
1.16.0 release.

As part of this release, the upstream now uses
upstream grpc, and thus some of the complexity
in the package can be dropped.

On this occasion, I have pulled a slightly newer
pubspec.yaml from the upstream that is yet to be
released which aided running the build with the
Flutter toolchain available in nixpkgs.