immich: 1.135.3 -> 1.136.0
`canvas` was added as a dependency to server/package.json in
https://github.com/immich-app/immich/commit/db0415bbcc7512e99013a6c310b64f51fd34a711

I am not sure if this was intentional as `canvas` is not ever imported
in the server package. Perhaps we can raise a bug for this.

https://github.com/immich-app/immich/releases/tag/v1.136.0

Note that the breaking change shouldn't affect NixOS users, as
`IMMICH_MEDIA_LOCATION` is set to an absolute path by default.

Signed-off-by: Sefa Eyeoglu <contact@scrumplex.net>