lixPackageSets.*.lix: only apply __structuredAttrs when needed
Commit 24cd08426a07616f036b270643b0538e44b37123 started building the lix
packages with __structuredAttrs as it is now required when separateDebugInfo
is combined with disallowedReferences. This breaks the static build by failing
to find toml11 with strange CMake errors [1]. Until the issue is fixed [2],
hack around this issue by only building with __structuredAttrs when separate
debug info is requested, which is not true of the static build.

[1] https://github.com/NixOS/nixpkgs/pull/424775#issuecomment-3092215936
[2] https://github.com/NixOS/nixpkgs/pull/428430