tic-80: add support for darwin
Already built 'successfully' with `NIXPKGS_ALLOW_UNSUPPORTED_SYSTEM=1`,
but the default CMake `installPhase` didn't do anything on macOS, so it
just resulted in an empty directory.

Bundle creation commands are based on [the ones in the README][], but
without the wrapper script since the extra args don't seem to be needed.

[the ones in the README]: https://github.com/nesbox/TIC-80#mac