grub2: refactor platforms logic, avoid abusing `meta.broken`
Previously, we used `platforms.gnu` and `platforms.linux` for Grub2's available
platforms, but blocking unsupported platforms via asserts. An attempt has been
made through marking unsupported platforms "broken" for EFI, but it would be
better if we just document the supported platforms in `meta.platforms` in the
first place. Also, the Grub2 package already has platform support for
PC/EFI/Xen, which deserves to be restructured in a unified way.

Regarding `meta.broken`, it shouldn't be used on invalid combinations of
arguments, as it will cause misunderstandings.

Changes have been made:
- Remove `meta.broken`, along with the 3 conditions in it.
- Remove `canEfi`, chain `efiSystemsBuild` on `meta.platform` if
  `pkgs.grub2_efi` is asked (building with `efiSupport = true;`)
- Avoid reusing `efiSystemsBuild` for Xen (PV), build a dedicated
  `xenSystemsBuild` for Xen, and chain that into `meta.platforms` in the same
  format.
- Revert the other 2 conditions (building with invalid arguments) back to
  asserts.

Signed-off-by: Hongbo <hehongbo@mail.com>