treewide: fix inappropriate usage of rust.envVars.setEnv
It was designed for the Cargo build hook, which goes into `nativeBuildInputs`.
Direct usage leads to an incorrect platform offset, which can for example be
observed by the package `pkgsStatic.buildPackages` differing from the native
one due to accidentally receiving the wrong environment variables. Fix the
issue by using `buildPackages` as `nativeBuildInputs` would.