cracklib: 2.10.0 -> 2.10.3
Use fetchFromGitHub to make the updateScript easier and avoid xz style
supply chain issues.

Separate into multiple outputs. The "out" output now doesn't have a bash
dependency anymore because all the scripts are in "bin".

Do not build the massive wordlist anymore. Instead rely on the built-in
one. This is also what e.g. archlinux does and it leads to a massively
smaller package: 10MiB -> 1MiB. This is especially dramatic/relevant
because cracklib is in the mandatory NixOS system closure because it's a
dependency of systemd. This change saves 9MiB transferred to literally
every NixOS machine.