sideswap: init updateScript
Automate updating of this package.

Added updateScript, which updates the package itself, including pubspec
and gitHashes files, as well as libsideswap-client sub-package, which should
be kept in sync with the main package. The version of the sub-package is taken
from sideswap/deploy/build_linux.sh file.

To update the package, run:
$ nix-shell maintainers/scripts/update.nix --argstr package sideswap