owl-lisp: Use dash for tests
On aarch64-linux the test fail to terminate, with bash printing a
message like this over and over again:

tests/run: line 63: wait: pid 105 is not a child of this shell

Running in GDB reveals that the is stuck in an infinite loop in
wait_for_background_pids waiting for a child that is no longer there,
suggesting that it's a bug in Bash.

Work around it for now by running tests in DASH instead.