shadow: use /bin/sh as default shell
See the comment in the code for the reason why using /bin/sh instead of
the runtimeShell store path.

Drop the patch and substituteInPlace in postPatch as this is much easier
to maintain than a patch.

Ldap users should now set `loginShell` to
`/run/current-system/sw/bin/bash` if they want to retain bash as the
default shell. If this is not set, the default shell for logins will be
/bin/sh which can be "turned" into a normal interactive bash by typing
"bash".