libpisp: fix backend.json path not found
Libpisp does a RUNPATH or RPATH lookup to check where the binary is installed and points to the incorrect location for the backend Json file. One way to solve this is to make the program  think it's not installed in which case, it will go to the correct backend.json file. For this we need to stop relying on RUNPATH, so we remove it.