libpthread-stubs: set platforms back to unix
I think this was changed from platforms.unix due to a misunderstanding
of what the README says.  The list in the README is of platforms where
the .pc file installed by this package will have one behaviour.  On
other Unix platforms, it will have a different behaviour, but it's
still supported.  Restoring platforms.unix fixes building a number of
packages for FreeBSD.

Fixes: 4cc23fda05f8 ("libpthread-stubs: refactored, moved to pkgs/by-name and renamed from xorg.libpthreadstubs")