odin: Fix arm64-apple-darwin builds
When installed on Apple Silicon, the Odin compiler will emit a warning
everytime it runs a build.

The warning is that `supplying the --target arm64-apple-macosx !=
arm64-apple-darwin argument to a nix-wrapped`.

This is because the Odin compiler uses the target name "arm64-apple-macosx"
while the nix-wrapped C compiler uses "arm64-apple-darwin".

The warning is emitted here:

https://github.com/NixOS/nixpkgs/blob/6770f4556182b1c83bb583eadaeef41902b30d53/pkgs/build-support/cc-wrapper/add-clang-cc-cflags-before.sh#L30

This patch fixes the warning by changing the target name used by the
Odin compiler to match the nix-wrapped C compiler.