llvmPackages: drop libclc attributes that error
libclc requires a compatible version of spirv-llvm-translator.  If we
don't have such a version, we get an eval error.  This means that
OfBorg builds of llvmPackages sets where libclc is unavailable will
produce an eval error, which is very confusing.  To fix this, restrict
the libclc attribute to LLVM versions where it actually exists.

Link: https://github.com/NixOS/nixpkgs/pull/414615#issuecomment-3126935312