util-linux: apply patch unconditionally
This was applied conditionally on staging-next to avoid throwing away
Linux builds.  It should have been changed to unconditional on
staging after that was merged, but I guess I forgot.  We generally
avoid conditional patch application wherever possible as it makes it
far too easy to miss when a patch needs to be removed or updated.