models-dev: add patch to rename hashed index file back to index.html
This change addresses an issue in Bun 1.2.13 (used in release 25.05)
where HTML entry points are generated with a hash appended to the
filename. As a result, the build fails unless the hashed index file is
renamed back to `index.html`.

The added patch scans the `./dist` directory for the hashed index
file and renames it back to `index.html` to ensure compatibility with
subsequent build steps.