buildMix: default to removing target config
See previous discussion at https://github.com/NixOS/nixpkgs/pull/429770.
cc @adamcstephens.

This is equivalent to giving `appConfigPath` an empty directory, but
expresses intent better (and doesn't require throwing an empty directory
into the store).

The Elixir ecosystem assumes [1] that dependencies are compiled without
their config; the `config/` directory is therefore used in library-only
projects to supply config values only intended for use when _developing_
them.  This leads to errors only seen in Nix when compile-time config
lacks runtime equivalents in end-user applications (per the whole
conversation at [1]).

Right now, the only way to get `buildMix` to build without config is to
manually remove the target's config directory in a hook/override, or (as
above) give `appConfigPath` an empty directory.  This PR adds a clearer
mechanism, and makes it the default.

[1] https://github.com/dashbitco/lazy_html/pull/11#issuecomment-3138715485

Tested with a `mix2nix`-using package that this functions as expected,
and with some `deps_nix`-using ones that `appConfigPath` still does too.