buildMix: add support for removing target config
This is equivalent to giving appConfigPath an empty directory, but
expresses intent better (and doesn't require throwing an empty directory
into the store).

The Elixir ecosystem assumes [1] that dependencies are compiled without
their config; the `config/` directory is therefore used in library-only
projects to supply config values only intended for use when _developing_
them.  This leads to errors only seen in Nix when compile-time config
lacks runtime equivalents in end-user applications (per the whole
conversation at [1]).

Right now, the only way to get buildMix to build without config is to
manually remove the target's config directory in a hook/override, or (as
above) give appConfigPath an empty directory.  This PR adds a clearer
mechanism, which would be nice to use from mix2nix et al.

[1] https://github.com/dashbitco/lazy_html/pull/11#issuecomment-3138715485