Revert "perf: inherit makeFlags from kernel"
This reverts commit 94276da9e98691be7d083b06301c3e434ba2cfa4.

Original justification[1] was:

> it reduces unnecessary code duplication and should help with cross
> compilation (although this cannot be tested yet because some
> dependencies fail).

However, as time as gone on its become increasingly clear that the
kernels make flags are going to be specific to building a kernel, not
a userspace program, hence them having been renamed to
kernelModuleMakeFlags in the meantime.  This has become even more
apparent now that we're using unwrapped compilers for the kernel,
which doesn't go at all well with building a userspace program like
perf, so we stop inheriting any make flags from the kernel, and
instead just set the two flags necessary for cross compilation to
work.

Link: https://github.com/NixOS/nixpkgs/pull/34068 [1]
Fixes: e9b2edb49163 ("linux: un-simplify toolchain selection")