haskell.packages.ghc9{2,4,6,8}.os-string: use 2.0.7
This matches what Stackage LTS 23 and thus we were using with GHC < 9.10
before: https://www.stackage.org/lts-23/package/os-string

This is actually not ideal since these GHC versions also ship versions
of filepath that may have module collisions with os-string, so using
os-string-1.0.0 (like for GHC < 9.2) would be better in some situations.
Unfortunately, it looks like the OsPath/os-string transition wasn't
completely thought through and e.g. hashable-1.4.7.0 (which we were also
using with Stackage LTS 23) has a hard dependency on os-string for
GHC >= 9.2, so we don't have much of a choice here.

The combination of a dependency on filepath and os-string doesn't seem
to be super common thus far, but down the line we may have trouble
getting such packages to work with GHC < 9.10 && >= 9.2.