zoom-us: reintroduce `unset QT_PLUGIN_PATH`
This was introduced long ago here

https://github.com/NixOS/nixpkgs/pull/116355
https://github.com/NixOS/nixpkgs/commit/b747aa0f9fd6363975e90a82210275a12bfbb696

to fix a display problem (participants dialog box).

With one of the more recent updates,
the workaround was no longer needed to fix
the original problem, so it was removed here

https://github.com/NixOS/nixpkgs/pull/425420
https://github.com/NixOS/nixpkgs/commit/c3a86691ba8aebe32cedbfc72e3b738f2bf93a1f

However, it turned out the workaround is still needed
to keep zoom happy in other environments, cf.

https://github.com/NixOS/nixpkgs/issues/428434
https://github.com/NixOS/nixpkgs/issues/429341