import-cargo-lock: optimize eval performance when allowBuiltinFetchGit
Motivation: Eval performance was very poor on some projects when using allowBuiltinFetchGit.

Nix seems to do some expensive operation on teh FS every time `builtins.fetchGit` is called, even when it is called with the same arguments multiple times.

This leads to a massive performance overhead if a cargo lock contains many dependencies referring to the same git url.

This change introduces a workaround for the issue, by deduplicating all the possible git fetcher calls to be done.