llvmPackages.tblgen: only install the tools we use
Otherwise we install a weird `update_core_linalg_named_ops.sh`
maintainer script that refers to the build directory.

There are explicit install targets for some of these and probably
better ways we could install the others, but the Clang derivation
also just explicitly copies a bunch of them, so meh.