salt: Build with the cryptography module
Without this, both the master and minion give errors like "NameError:
name 'rsa' is not defined", while with it they both run fine. This is
presumably because Salt just ignores import errors of the cryptography
library[1], so rather than erroring immediately on startup it takes a while
(at least for the minion).

[1] https://github.com/saltstack/salt/blob/master/salt/crypt.py#L50