davinci-resolve: only add updateScript to non-studio
This is an attempt to “fix” automated updates having the commit message
prefixed with `davinci-resolve-studio:`. Now they should always be
`davinci-resolve:`. In theory it should be `davinci-resolve{,-studio}:`,
but this seems fine (and is realistically how we prefix human-made ones
anyway). Idea here is that automated update will only ever happen for
the non-studio (which updates both).