paretosecurity: 0.2.38 -> 0.3.2
Test for linking had to be removed because now the linking is more
complex and it would take too much mocking to test it.

The test in question was moved to the dashboard codebase that is
closed-source, if that helps in any way.

(cherry picked from commit a667834a5fe2e18dd4405ddc19679dae7566c4b1)