nixos/{consul-template,vault-agent}: drop `template` sub-option
An HCL config can be represented as JSON in different ways and we forced
users to use a list of attrsets for `cfg.settings.template` prior to
this change.

Similarly, a user may opt to use `vault-agent` exclusively as auth-proxy
without any templates, which would result in `null` ending up in the
resulting JSON configuration and prevent `vault-agent` from starting.

A lot of NixOS modules opt to recursively filter out any `null` values
when hitting this, either by defining a function in `apply =` in
`cfg.settings` or in the `format.generate` call, but that comes with its
own set of downsides.

While I personally can't think of a use-case involving `consul-template`
 without templates, it does manage to start without one just fine and
`consul-template` is equally affected by our opinionated JSON
representation forced onto the user.

As such, we decided in favor of removing the option entirely in the
discussion leading up to this.

This is one of the 3 suggested implementations from issue 420208.
You can see the other 2 there too, if you are curious.

tl;dr: Non-breaking bug fix to allow using `vault-agent` without
templates and no longer forcing our opinionated JSON representation onto
the user.