ycmd: unstable-2023-11-06 -> unstable-2025-06-16
Update `version`, `rev`, and `sha256` because why not update if also
fixing dependency renaming stuff.

Add `legacy-cgi` due to breaking changes in Python updates;

- https://pypi.org/project/legacy-cgi/
- https://peps.python.org/pep-0594/#deprecated-modules
- https://peps.python.org/pep-0594/#cgi

TLDR: `cgi` was deprecated in version `3.11` and aimed to be removed in
`3.12` of Python due to;

> “[...] designed poorly and are now near-impossible to fix (`cgi`)
> [...]”

Remove `disable` attribute at request of @SuperSandro2000 because
Python2 is EOL

Add symbolic link to address Issue #429485 with YouCompleteMe Vim
plugin, however, beware there are _opportunities_ for future bug reports
to be opened because of `ycmd` and YouCompleteMe version de-sync.

Add `longDescription` with notes about YouCompleteMe errors when
run-time and compile-time Python interpreters do not match.