nixos/homepage-dashboard: add support for `proxmox` config
The upstream project recently added support for
a new configuration file named `proxmox.yaml`.

Without adding this, the server can fail to start
since it tries to create the file `proxmox.yaml`
but fails because `/etc` is read-only.

See https://gethomepage.dev/configs/proxmox/ for
details of the feature, which is handled here
similiarly to docker/kubernetes configs.