haskell.packages.ghc{810,90,92}.ghc-internal: add missing attribute
See 1b5377100b83e18d0b6b4e18fe2dadbadaf98e0f. This is required to
evaluate some packages since it is an argument to their package
expression. We can't actually provide a version that works with these
GHC versions, so we set it to null and hope the packages themselves only
have a conditional dependency on them.