haskell.packages.*.ghc-{toolchain,platform}: make sure attr exists
These packages are new core packages starting with GHC 9.10, but there
is no (simple) way to make them available for older GHCs. We just need
to make sure the attributes exist, so we can evaluate packages that
depend on them with all GHC versions (even if the build may fail).