lkl: fix hijack and zpoline when building in parallel
because they both share object files, which may result in symbols
missing from the resulting build outputs.

E.g. `liblkl-hijack.so` is used by `nftables-check` and would fail
because of missing symbols caused by this:

~~~
/nix/store/x6mqw6qfcmya8jvhrmk938zirvqb6hhr-nftables-1.1.3/bin/nft: symbol lookup error: /nix/store/fgzk4jdjhmz86d8vmgdjg6l42gbmfa1m-lkl-2025-03-20-lib/lib/liblkl-hijack.so: undefined symbol: dual_fds
~~~

Note: The patch is vendored, because upstream has not merged this PR yet
due to Makefile compatibility concerns with older distros.

Co-authored-by: Raito Bezarius <masterancpp@gmail.com>