ollama: disable failing tests on darwin
Only some tests fail, but it seems we would want to skip testing a whole
subpackage and that is not doable in a flag to `go test` and
`buildGoModule` would have to be modified (or duplicated to this drv).
Seems better to disable tests for now.

See https://github.com/NixOS/nixpkgs/issues/431464.