kanidm: use `finalAttrs` everywhere
This makes the resulting derivation when using `overrideAttrs` more
accurate and also fixes a discrepancy in `kanidmWithSecretProvisioning`.

One would expect this change to be a no-op, but it isn't. Not because
there is some unwanted change, but rather because unlike
`kanidm_1_x.withSecretProvisioning`, `kanidmWithSecretProvisioning_1_x`
used `kanidm` as a basis when it should have used `kanidm_1_x`.

~~~
# before
nix-repl> kanidmWithSecretProvisioning_1_5.patches
[
  ./pkgs/by-name/ka/kanidm/provision-patches/1_6/oauth2-basic-secret-modify.patch
  ./pkgs/by-name/ka/kanidm/provision-patches/1_6/recover-account.patch
]

# after
nix-repl> kanidmWithSecretProvisioning_1_5.patches
[
  ./pkgs/by-name/ka/kanidm/provision-patches/1_5/oauth2-basic-secret-modify.patch
  ./pkgs/by-name/ka/kanidm/provision-patches/1_5/recover-account.patch
]
~~~

The change itself is written in a way that minimizes formatting changes
enforced by `nixfmt`. For example, it would have been nice to access
`finalAttrs.version` in `versionUnderscored` without having to make it a
function, but that would have required to move the `let in` into
`finalAttrs: {`, causing the entire attrset to reformat. Similarly,
offloading this into `passthru` would have caused other bits to reformat
because of the variable name length differences.